Notify adapter dataset change in onPause()
notifyDataSetChanged() should still be called even in onPause() method of
activity since some UI events such as fling might still be processed
after onPause() is called. This patch also fixes the bug that adapter might
be null when project loading failed and should be checked for null when
doing the cleanup.

We also moved the original "onStart() and onStop()" setup/cleanup to
"onResume() and onPause()" to fix the bug 5090976. But we don't have a
clear explanation as to why it works. Without this change, the monkey
test still fails even though adapter is taken care of. We will keep the
bug open and investigate the reason later on to make sure this bug can
be closed safely. The bug could also manifest itself as we move on. For
now with this change, the monkey test passed.

Bug: 5091156, 5093407, 5090976
Change-Id: I1bb06e1d752960a94c8f094060955c99fca84eb4