Fix 4601568: video playback pauses after orientation change.
There are several problems:

(1) We might not have a chance restart playing before saving the playing
state and finishing (because we need to wait for surfaceCreated()). So
when saving the playing state, we need to check mRestartPreview as well.

(2) We still need to stop preview even if the activity is being shutdown.

(3) Post mProcessQueueRunnable when mThreadHandler is created, otherwise
the runnable may be posted after the thread itself is gone.

Change-Id: I44794ac40daec9376d238a63fbe254aeece45f10